Removed divs, and moved helpfunctions out
