+++ app/assets/main/react_components/dashboard/climate_actions/carousel/CarouselCard.jsx
@@ -55,50 +55,44 @@ const CarouselCard = ({
   }, []);
 
   return (
-    <div className="flex flex-1 min-h-full">
-      <div className="pt-20 flex m-lg:pt-24 flex-1 justify-evenly">
-        <div className=" border-gray-tint-2 rounded-lg shadow-lg pb-2 ml-2 mr-2 flex flex-col flex-1 bg-white">
-          <CategoryColorBanner categories={categories} action={action} />
-          <div
-            className={`mx-auto bg-gray-tint-2 bg-opacity-10 shadow-md -mt-24 rounded-full h-40 w-40 items-center justify-center bg-cover filter drop-shadow-xl`}
-            style={{
-              backgroundImage: action.image_url
-                ? `url('${action.image_url}')`
-                : "url('/action_images/Globe.png')",
-              backgroundSize: "100%",
-            }}
-          >
-            {action.action_of_the_month && !monthlyActionBanner && (
-              <TextBanner text={text} />
+    <div className="min-h-full pt-20 flex m-lg:pt-24 justify-evenly">
+      <div className="border-gray-tint-2 rounded-lg shadow-lg pb-2 mx-2 flex flex-col bg-white">
+        <CategoryColorBanner categories={categories} action={action} />
+        <div
+          className={`mx-auto bg-gray-tint-2 bg-opacity-10 shadow-md -mt-24 rounded-full h-40 w-40 items-center justify-center bg-cover filter drop-shadow-xl`}
+          style={{
+            backgroundImage: action.image_url
+              ? `url('${action.image_url}')`
+              : "url('/action_images/Globe.png')",
+            backgroundSize: "100%",
+          }}
+        >
+          {action.action_of_the_month && !monthlyActionBanner && (
+            <TextBanner text={text} />
+          )}
+        </div>
+        <div className="flex flex-col flex-1 text-center mx-2">
+          <ImpactPoints action={action} />
+          <h3 className="flex-1 justify-center align-center text-base font-bold self-center">
+            {action.name.length > 40
+              ? action.name.slice(0, 40) + "..."
+              : action.name}
+          </h3>
+          <p className="flex-4 text-sm">
+            {action.description.length > 200
+              ? action.description.slice(0, 200) + "..."
+              : action.description}
+          </p>
+          <div className="flex-1 mt-5 justify-center align-center">
+            <AcceptanceStatistics action={action} />
+            {!action.accepted && (
+              <button
+                className="button inline-block "
+                onClick={() => handleClickAccepted(action)}
+              >
+                {climateActionsText.accept}
+              </button>
             )}
-          </div>
-          <div className="flex flex-col flex-1 text-center mx-2">
-            <ImpactPoints action={action} />
-            <div className="flex-1 justify-center align-center self-center">
-              <h3 className={`text-base font-bold self-center`}>
-                {action.name.length > 40
-                  ? action.name.slice(0, 40) + "..."
-                  : action.name}
-              </h3>
-            </div>
-            <div className="flex-4">
-              <p className="text-sm">
-                {action.description.length > 200
-                  ? action.description.slice(0, 200) + "..."
-                  : action.description}
-              </p>
-            </div>
-            <div className="flex-1 mt-5 justify-center align-center">
-              <AcceptanceStatistics action={action} />
-              {!action.accepted && (
-                <button
-                  className="button inline-block "
-                  onClick={() => handleClickAccepted(action)}
-                >
-                  {climateActionsText.accept}
-                </button>
-              )}
-            </div>
           </div>
         </div>
       </div>
