+++ app/assets/main/react_components/dashboard/climate_actions/carousel/CarouselHeader.jsx
@@ -3,16 +3,13 @@ import { useClimateActionsText } from "../../../contexts/TextContext.js";
 
 const CarouselHeader = () => {
   const climateActionsText = useClimateActionsText();
-
   return (
-    <div>
+    <>
       <h3 className="heading-lg mb-3">{climateActionsText.take_action}</h3>
-
       <h3 className="heading mb-3">
         {climateActionsText.low_effort_high_impact_actions}
       </h3>
-      <div className="flex flex-col t:flex-row t:space-x-8 space-y-6 t:space-y-0"></div>
-    </div>
+    </>
   );
 };
