+++ app/assets/main/react_components/dashboard/climate_actions/carousel/carousel_card_attributes/ImpactPoints.jsx
@@ -2,11 +2,11 @@ import React from "react";
 
 const ImpactPoints = ({ action }) => {
   return (
-    <div className="flex-1 flex flex-row justify-center self-center">
+    <div className="flex flex-row justify-center self-center">
       {[1, 2, 3, 4, 5].map((index) => {
         return (
           <span
-            className={`flex flex-row self-center bg-gray-tint-2 m-2 rounded-full h-6 w-6 justify-center bg-cover`}
+            className={`self-center bg-gray-tint-2 m-2 rounded-full h-6 w-6 justify-center bg-cover`}
             key={action.name + index}
             style={
               index <= action.points
@@ -15,7 +15,7 @@ const ImpactPoints = ({ action }) => {
                   }
                 : {}
             }
-          ></span>
+          />
         );
       })}
     </div>
