+++ app/assets/main/react_components/dashboard/climate_actions/manage_actions/kanban/KanbanActionContainer.jsx
@@ -2,17 +2,17 @@ import React, { useState, useRef, useEffect } from "react";
 import { DragDropContext } from "react-beautiful-dnd";
 import KanbanActionColumn from "./KanbanActionColumn.jsx";
 import { useDeletedActionUpdate } from "../../../../contexts/DeletedActionContext.js";
-import { orderBy } from "lodash";
 import {
   useUserState,
   useUserActions,
 } from "../../../../contexts/UserContext.js";
-import {
-  updateStatus,
-  deleteUserAction,
-} from "../../../../helpers/DBRequests.js";
+import { deleteUserAction } from "../../../../helpers/DBRequests.js";
 import { useMediaQuery } from "react-responsive";
 import { t } from "../../../../constants";
+import {
+  onDragEnd,
+  collectPerformedUserActions,
+} from "../../../../helpers/KanbanHelper.js";
 
 const KanbanActionContainer = ({
   sidebarCollapsed,
@@ -79,170 +79,6 @@ const KanbanActionContainer = ({
     setDeletedAction(deletedAction);
   };
 
-  const collectPerformedUserActions = (destItems) => {
-    let performedUserActions = [];
-    destItems.map((category) => {
-      return category.userActionsArray.map((userAction) => {
-        if (userAction.status === true) {
-          userAction.id.toString();
-          performedUserActions = [...performedUserActions, userAction];
-        }
-      });
-    });
-    return performedUserActions;
-  };
-
-  const sortActionsBasedOnStatus = (performedActions) => {
-    return orderBy(performedActions, ["status"], ["desc"]);
-  };
-
-  const handleCompleteAction = (movedItem) => {
-    //Move items in kanban through buttons instead of drag and drop
-    //Button for performing an action
-    const sourceColumn = columns[1];
-    const destColumn = columns[2];
-    const sourceItems = [...sourceColumn.items];
-    //Updade status both locally and DB needs to happen before the filtering through status below
-    updateStatus(movedItem.id, true);
-    movedItem.status = true;
-    //Filter out moved item from first column by local status.
-    const filteredSourceItems = sourceItems.filter(
-      (item) => item.status === false
-    );
-    //Add moved item to second column. Helpfunction in context desides if new categoryBadge should be created or just change status and color on subitem
-    const destItems = updateAchievementsOnMove(movedItem, destColumn.items);
-    const sortedDestItems = destItems.map((category) => {
-      return {
-        ...category,
-        userActionsArray: sortActionsBasedOnStatus(category.userActionsArray),
-      };
-    });
-    updateAchievements([...sortedDestItems]);
-    const newSortedDestItems = orderBadgesOnItemDragged(
-      sortedDestItems,
-      movedItem
-    );
-    //Function to get performed useraction from categoryBadges
-    let performedUserActions = collectPerformedUserActions(newSortedDestItems);
-    updateUserActions([...sourceItems, ...performedUserActions]);
-
-    updateColumns({
-      ...columns,
-      [1]: {
-        ...sourceColumn,
-        items: filteredSourceItems,
-      },
-      [2]: {
-        ...destColumn,
-        items: newSortedDestItems,
-      },
-    });
-  };
-
-  const handleUncompleteAction = (movedItem) => {
-    /** Button for unperform */
-    const sourceColumn = columns[2];
-    const destColumn = columns[1];
-    const sourceItems = [...sourceColumn.items];
-    const destItems = [...destColumn.items];
-    //When we unperform a subitem within a categoryBadge we "create" a new userAction card, which needs an id as string
-    movedItem.id = movedItem.id.toString();
-    //We add the card in the end of the list
-    const destIndex = destItems.length;
-    destItems.splice(destIndex, 0, movedItem);
-    //change the local status and in DB
-    movedItem.status = false;
-    updateStatus(movedItem.id, false);
-    //Change the status of the subitem so that color can depend the categoryBadge
-    const newSourceItems = sourceItems.map((category) => {
-      return {
-        ...category,
-        userActionsArray: category.userActionsArray.map((item) => {
-          return item.id == movedItem.id ? { ...item, status: false } : item;
-        }),
-      };
-    });
-
-    const sortedSourceItems = newSourceItems.map((category) => {
-      return {
-        ...category,
-        userActionsArray: sortActionsBasedOnStatus(category.userActionsArray),
-      };
-    });
-    //Check if categoryBadge has any subitem with status true, else delete the categoryBadge
-    const checkDelete = sortedSourceItems.filter((category) => {
-      return category.userActionsArray.some((item) => item.status === true);
-    });
-    updateAchievements([...checkDelete]);
-    updateUserActions([...destItems]);
-    updateColumns({
-      ...columns,
-      [1]: {
-        ...destColumn,
-        items: destItems,
-      },
-      [2]: {
-        ...sourceColumn,
-        items: checkDelete,
-      },
-    });
-  };
-
-  const orderBadgesOnItemDragged = (performedCategories, movedItem) => {
-    return orderBy(performedCategories, ({ id }) =>
-      id == movedItem.climate_action_category_id ? 0 : 1
-    );
-  };
-
-  const onDragEnd = (result, columns) => {
-    if (!result.destination) return;
-    const { source, destination } = result;
-    if (source.droppableId !== destination.droppableId) {
-      const sourceColumn = columns[source.droppableId];
-      const destColumn = columns[destination.droppableId];
-      const sourceItems = [...sourceColumn.items];
-      const [removed] = sourceItems.splice(source.index, 1);
-      updateStatus(removed.id, true, mounted);
-      const destItems = updateAchievementsOnMove(removed, destColumn.items);
-      const sortedDestItems = destItems.map((category) => {
-        return {
-          ...category,
-          userActionsArray: sortActionsBasedOnStatus(category.userActionsArray),
-        };
-      });
-      let performedUserActions = collectPerformedUserActions(destItems);
-      updateUserActions([...sourceItems, ...performedUserActions]);
-      const newSortedDestItems = orderBadgesOnItemDragged(
-        sortedDestItems,
-        removed
-      );
-      updateAchievements([...newSortedDestItems]);
-      updateColumns({
-        ...columns,
-        [source.droppableId]: {
-          ...sourceColumn,
-          items: sourceItems,
-        },
-        [destination.droppableId]: {
-          ...destColumn,
-          items: newSortedDestItems,
-        },
-      });
-    } else {
-      const column = columns[source.droppableId];
-      const copiedItems = [...column.items];
-      const [removed] = copiedItems.splice(source.index, 1);
-      copiedItems.splice(destination.index, 0, removed);
-      updateColumns({
-        ...columns,
-        [source.droppableId]: {
-          ...column,
-          items: copiedItems,
-        },
-      });
-    }
-  };
-
   useEffect(() => {
     mounted.current = true;
     return () => {
@@ -252,7 +88,19 @@ const KanbanActionContainer = ({
 
   return (
     <div className="h-screen">
-      <DragDropContext onDragEnd={(result) => onDragEnd(result, columns)}>
+      <DragDropContext
+        onDragEnd={(result) =>
+          onDragEnd(
+            result,
+            mounted,
+            columns,
+            updateUserActions,
+            updateColumns,
+            updateAchievements,
+            updateAchievementsOnMove
+          )
+        }
+      >
         {Object.entries(columns).map(([columnId, column]) => {
           return (
             <div
@@ -278,8 +126,6 @@ const KanbanActionContainer = ({
                 columnId={columnId}
                 key={columnId}
                 handleDelete={handleDelete}
-                handleCompleteAction={handleCompleteAction}
-                handleUncompleteAction={handleUncompleteAction}
                 categories={categories}
                 setSidebarCollapsed={setSidebarCollapsed}
                 sidebarCollapsed={sidebarCollapsed}
