+++ app/assets/main/react_components/dashboard/climate_actions/manage_actions/kanban/card_components/UserActionCard.jsx
@@ -6,7 +6,6 @@ const UserActionCard = ({
   categoryColor,
   userAction,
   handleDelete,
-  handleCompleteAction,
   sidebarCollapsed,
 }) => {
   return (
@@ -19,23 +18,18 @@ const UserActionCard = ({
             "_active"
           } h-7 w-full rounded-t border-t-gray-tint-2 bg-opacity-60 top-0 absolute z-0`}
         ></div>
-
         <KanbanCardImage
           img={userAction.image_url}
           sidebarCollapsed={sidebarCollapsed}
           isUserAction={true}
         ></KanbanCardImage>
       </div>
-
-      <div>
-        {userAction.expanded && (
-          <UserActionCardExpanded
-            userAction={userAction}
-            handleDelete={handleDelete}
-            handleCompleteAction={handleCompleteAction}
-          ></UserActionCardExpanded>
-        )}
-      </div>
+      {userAction.expanded && (
+        <UserActionCardExpanded
+          userAction={userAction}
+          handleDelete={handleDelete}
+        ></UserActionCardExpanded>
+      )}
     </div>
   );
 };
